Move asset view only in dashboard view
